requirements.txt
python-telegram-bot[job-queue]>=20.0
httpx>=0.25.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
cryptography>=41.0.0
